Break tests out into seperate module files.
# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]
