+++ src/rust/engine/client/src/options/id.rs
@@ -123,58 +123,3 @@ impl OptionId {
       .join(sep)
   }
 }
-
-#[cfg(test)]
-mod test {
-  use crate::options::id::{OptionId, Scope};
-
-  #[test]
-  fn test_option_id_global_switch() {
-    let option_id = option_id!(-'x', "bar", "baz");
-    assert_eq!(
-      OptionId::new(Scope::GLOBAL, ["bar", "baz"].iter(), Some('x')).unwrap(),
-      option_id
-    );
-    assert_eq!("GLOBAL", option_id.scope());
-  }
-
-  #[test]
-  fn test_option_id_global() {
-    let option_id = option_id!("bar", "baz");
-    assert_eq!(
-      OptionId::new(Scope::GLOBAL, ["bar", "baz"].iter(), None).unwrap(),
-      option_id
-    );
-    assert_eq!("GLOBAL", option_id.scope());
-  }
-
-  #[test]
-  fn test_option_id_scope_switch() {
-    let option_id = option_id!(-'f', ["foo-bar"], "baz", "spam");
-    assert_eq!(
-      OptionId::new(
-        Scope::Scope("foo-bar".to_owned()),
-        ["baz", "spam"].iter(),
-        Some('f')
-      )
-      .unwrap(),
-      option_id
-    );
-    assert_eq!("foo-bar", option_id.scope());
-  }
-
-  #[test]
-  fn test_option_id_scope() {
-    let option_id = option_id!(["foo-bar"], "baz", "spam");
-    assert_eq!(
-      OptionId::new(
-        Scope::Scope("foo-bar".to_owned()),
-        ["baz", "spam"].iter(),
-        None
-      )
-      .unwrap(),
-      option_id
-    );
-    assert_eq!("foo-bar", option_id.scope());
-  }
-}
