+++ src/rust/engine/client/src/pantsd.rs
@@ -8,12 +8,12 @@ use libc::pid_t;
 use log::debug;
 use sha2::{Digest, Sha256};
 
-struct Metadata {
+pub(crate) struct Metadata {
   metadata_dir: PathBuf,
 }
 
 impl Metadata {
-  fn mount<P: AsRef<Path>>(directory: P) -> Result<Metadata, String> {
+  pub(crate) fn mount<P: AsRef<Path>>(directory: P) -> Result<Metadata, String> {
     let info = uname::uname().map_err(|e| format!("{}", e))?;
     let host_hash = Sha256::new()
       .chain(&info.sysname)
@@ -74,7 +74,7 @@ impl Metadata {
     self.read_metadata("process_name").map(|(_, value)| value)
   }
 
-  fn port(&self) -> Result<u16, String> {
+  pub(crate) fn port(&self) -> Result<u16, String> {
     self
       .read_metadata("socket")
       .and_then(|(socket_metadata_path, value)| {
@@ -172,86 +172,5 @@ pub fn probe(working_dir: &Path, metadata_dir: &Path) -> Result<u16, String> {
       expected_name = expected_process_name_prefix,
       actual_name = actual_argv0
     ))
-  }
-}
-
-#[cfg(test)]
-mod test {
-  use std::fs;
-  use std::net::TcpStream;
-  use std::process::{Command, Stdio};
-  use std::str::from_utf8;
-
-  use tempdir::TempDir;
-
-  use crate::build_root::BuildRoot;
-  use crate::pantsd;
-
-  fn launch_pantsd() -> (BuildRoot, TempDir) {
-    let build_root = BuildRoot::find()
-      .expect("Expected test to be run inside the Pants repo but no build root was detected.");
-    let pants_subprocessdir = TempDir::new("pants_subproccessdir").unwrap();
-    let mut cmd = Command::new(build_root.join("pants"));
-    cmd
-      .current_dir(build_root.as_path())
-      .arg("--pants-config-files=[]")
-      .arg("--no-pantsrc")
-      .arg("--pantsd")
-      .arg(format!(
-        "--pants-subprocessdir={}",
-        pants_subprocessdir.path().display()
-      ))
-      .arg("-V")
-      .stderr(Stdio::inherit());
-    let result = cmd
-      .output()
-      .map_err(|e| {
-        format!(
-          "Problem running command {command:?}: {err}",
-          command = cmd,
-          err = e
-        )
-      })
-      .unwrap();
-    assert_eq!(Some(0), result.status.code());
-    assert_eq!(
-      fs::read_to_string(
-        build_root
-          .join("src")
-          .join("python")
-          .join("pants")
-          .join("VERSION")
-      )
-      .unwrap(),
-      from_utf8(result.stdout.as_slice()).unwrap()
-    );
-    (build_root, pants_subprocessdir)
-  }
-
-  fn assert_connect(port: u16) {
-    assert!(
-      port >= 1024,
-      "Pantsd should never be running on a privileged port."
-    );
-
-    let stream = TcpStream::connect(("0.0.0.0", port)).unwrap();
-    assert_eq!(port, stream.peer_addr().unwrap().port());
-  }
-
-  #[test]
-  fn test_address_integration() {
-    let (_, pants_subprocessdir) = launch_pantsd();
-
-    let pantsd_metadata = pantsd::Metadata::mount(&pants_subprocessdir).unwrap();
-    let port = pantsd_metadata.port().unwrap();
-    assert_connect(port);
-  }
-
-  #[test]
-  fn test_probe() {
-    let (build_root, pants_subprocessdir) = launch_pantsd();
-
-    let port = pantsd::probe(&build_root, pants_subprocessdir.path()).unwrap();
-    assert_connect(port);
   }
 }
